delete product / category / brand images with deleteone function

+++ services/handlersFactory.js
@@ -5,15 +5,57 @@ const fs = require("fs");
 exports.deleteOne = (Model) =>
   asyncHandler(async (req, res, next) => {
     const oldProduct = await Model.findOne({ _id: req.params.id });
-    const oldImage = oldProduct.image.split("http://localhost:8000/")[1];
-    fs.unlink("C://projects/ecommerce-apis/uploads/" + oldImage, (err) => {
-      if (err) {
-        // file doens't exist
-        console.info("File doesn't exist, won't remove it.");
-      } else {
-        console.info(`removed`);
-      }
-    });
+
+    //desc :  logic to delete product Images from the server.
+
+    if (oldProduct.images) {
+      //delete product images array
+      let imagesArr = [];
+      oldProduct.images.forEach((image) =>
+        imagesArr.push(image.split("http://localhost:8000/")[1])
+      );
+      imagesArr.forEach((image) =>
+        fs.unlink("C://projects/ecommerce-apis/uploads/" + image, (err) => {
+          if (err) {
+            // file doens't exist
+            console.info("File doesn't exist, won't remove it.");
+          } else {
+            console.info(`removed`);
+          }
+        })
+      );
+
+      // desc : delete product cover image
+      const oldCoverImage = oldProduct.imageCover.split(
+        "http://localhost:8000/"
+      )[1];
+      console.log(oldCoverImage);
+      fs.unlink(
+        "C://projects/ecommerce-apis/uploads/" + oldCoverImage,
+        (err) => {
+          if (err) {
+            // file doens't exist
+            console.info("File doesn't exist, won't remove it.");
+          } else {
+            console.info(`removed`);
+          }
+        }
+      );
+    }
+
+    //desc:  logic to delete brand / category image from th server
+    if (oldProduct.image) {
+      const oldImage = oldProduct.image.split("http://localhost:8000/")[1];
+      fs.unlink("C://projects/ecommerce-apis/uploads/" + oldImage, (err) => {
+        if (err) {
+          // file doens't exist
+          console.info("File doesn't exist, won't remove it.");
+        } else {
+          console.info(`removed`);
+        }
+      });
+    }
+
     const document = await Model.findByIdAndDelete(req.params.id);
     if (!document) {
       return next(
